Struct cpufreq::Cpu

pub struct Cpu { /* fields omitted */ }

Methods

impl Cpu

fn get_all() -> Iterator

Iterate over all Cpu's available in your system

fn exists(id: CpuId) -> bool

Check whether a Cpu with given ID exists in you system

fn new(id: CpuId) -> Cpu

fn get_freq(&self) -> Result<Frequency, CpuPowerError>

Get frequency reported by hardware or by kernel This function tries to get freq using call to hardware first, and if that call fails - uses call to kernel

fn get_freq_kernel(&self) -> Result<Frequency, CpuPowerError>

Get frequency reported by your kernel According to the underlying library documentation - you don't need to be root to perform this operation

fn get_freq_hardware(&self) -> Result<Frequency, CpuPowerError>

Get frequency reported by your hardware According to the underlying library documentation - you should be root to perform this operation

fn set_freq(&self, freq: Frequency) -> Result<&Cpu, CpuPowerError>

Set frequency for the given CPU You should have root privileges to do that

fn get_transition_latency(&self) -> Result<u64, CpuPowerError>

Determine CPUs transition latency Returns: transition latency in nanoseconds (10-9 s)

fn modify_policy_max(&self, max: Frequency) -> Result<(), CpuPowerError>

Modify current policy by changing it's max frequency

fn modify_policy_min(&self, min: Frequency) -> Result<(), CpuPowerError>

Modify current policy by changing it's min frequency

fn modify_policy_governor(&self, governor: &str) -> Result<(), CpuPowerError>

Modify current policy by changing it's governor

fn get_hardware_limits(&self) -> Result<(Frequency, Frequency), CpuPowerError>

Determine hardware CPU frequency limits

These may be limited further by thermal, energy or other considerations by cpufreq policy notifiers in the kernel.

fn get_id(&self) -> CpuId

Get if of the current processor

fn get_driver(&self) -> Result<String, CpuPowerError>

Determine CPUfreq driver used

fn get_policy(&self) -> Result<Policy, CpuPowerError>

Determine CPUfreq policy used

You can try to change current policy by using set_policy method

fn set_policy(&self, policy: &Policy) -> Result<(), CpuPowerError>

Set new CPUfreq policy to use This tries to set the passed policy as new policy as close as possible, but results may differ depending e.g. on governors being available.

fn get_available_governors(&self) -> Result<Vec<String>, CpuPowerError>

determine CPUfreq governors currently available

may be modified by modprobe'ing or rmmod'ing other governors

fn get_available_frequencies(&self) -> Result<Vec<Frequency>, CpuPowerError>

Get frequencies available for the given CPU

fn get_affected_cpus(&self) -> Result<Vec<Cpu>, CpuPowerError>

fn get_related_cpus(&self) -> Result<Vec<Cpu>, CpuPowerError>

fn get_stats(&self) -> Result<Vec<Stat>, CpuPowerError>

Determine stats for the cpufreq subsystem

fn get_transitions(&self) -> Result<u64, CpuPowerError>

Determine total transition count for this CPU